+++ src/main/java/ru/vyarus/dropwizard/orient/OrientServerBundle.java
@@ -2,11 +2,9 @@ package ru.vyarus.dropwizard.orient;
 
 import io.dropwizard.core.Configuration;
 import io.dropwizard.core.ConfiguredBundle;
-import io.dropwizard.jersey.validation.Validators;
 import io.dropwizard.core.setup.Bootstrap;
 import io.dropwizard.core.setup.Environment;
 import org.eclipse.jetty.servlet.ServletHolder;
-import org.hibernate.validator.internal.engine.resolver.JPATraversableResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.vyarus.dropwizard.orient.configuration.OrientServerConfiguration;
@@ -15,11 +13,9 @@ import ru.vyarus.dropwizard.orient.configuration.deserializer.NetworkProtocolDes
 import ru.vyarus.dropwizard.orient.configuration.deserializer.ParameterDeserializer;
 import ru.vyarus.dropwizard.orient.health.OrientServerHealthCheck;
 import ru.vyarus.dropwizard.orient.internal.EmbeddedOrientServer;
-import ru.vyarus.dropwizard.orient.internal.TraverseAllResolver;
 import ru.vyarus.dropwizard.orient.support.ConsoleCommand;
 import ru.vyarus.dropwizard.orient.support.OrientServlet;
 
-import javax.validation.TraversableResolver;
 import java.util.function.Function;
 
 /**
@@ -31,12 +27,6 @@ import java.util.function.Function;
  * Server startup could be disabled by setting 'start: false' in config.
  * <p>
  * Also, registers console command for interacting with database.
- * <p>
- * Important note: orient object database conflicts with hibernate-validator, because orient brings jpa api jar,
- * which activates hibernate lazy properties checks in validator. Bundle implicitly fixes this by overriding
- * {@link javax.validation.ValidatorFactory} in bootstrap. It should not bring any side effects for most cases, but
- * if you declare custom {@link javax.validation.ValidatorFactory} then simply declare {@link TraversableResolver}
- * manually in it to prevent override (e.g. into {@link TraverseAllResolver}).
  * <p>
  * NOTE: server will not start when console command called, because dropwizard will not run managed objects this time
  * (only server command triggers managed objects lifecycle). But plocal connections still could be used.
@@ -68,7 +58,6 @@ public class OrientServerBundle<T extends Configuration> implements ConfiguredBu
         bootstrap.getObjectMapper().addHandler(new ParameterDeserializer());
         bootstrap.getObjectMapper().addHandler(new NetworkProtocolDeserializer());
 
-        recoverValidatorBehaviour(bootstrap);
         bootstrap.addCommand(new ConsoleCommand(
                 bootstrap.getApplication().getConfigurationClass(),
                 configurationProvider));
@@ -93,52 +82,6 @@ public class OrientServerBundle<T extends Configuration> implements ConfiguredBu
         if (conf.isAdminServlet()) {
             environment.getAdminContext().addServlet(new ServletHolder(
                     new OrientServlet(orientServer.getServerInfo())), "/orient/*");
-        }
-    }
-
-    /**
-     * Orientdb-object module includes hiberate-jpa-api into classpath and hibernate-validator
-     * starts to think that JPA is available (if jpa available then validated objects must be checked with
-     * traversable provider to avoid lazy init exceptions).
-     * But orient's {@code OJPAPersistenceProvider.getProviderUtil()} simply throws exception.
-     * <p>
-     * The only way to resolve problem is to manually register correct {@link TraversableResolver}.
-     * (see {@link org.hibernate.validator.internal.engine.resolver.TraversableResolvers#getDefault()} for
-     * default resolver selection logic).
-     * <p>
-     * It's not a hack: in normal case, validator use the same implementation, so it's only behaviour fix due to
-     * additional jpa jar appeared in classpath (validator is wrong assuming it means jpa will be really used).
-     * This substitution could only override manually registered factory and to avoid this create this
-     * factory with correct resolver.
-     * <p>
-     * Note that it can't cause side effects because hibernate is actually not used.
-     */
-    private void recoverValidatorBehaviour(final Bootstrap<?> bootstrap) {
-        final TraversableResolver resolver = bootstrap.getValidatorFactory().getTraversableResolver();
-        if (isObjectOrientUsed() && resolver instanceof JPATraversableResolver) {
-            logger.debug("Overriding incorrectly configured ValidatorFactory to remove jpa support");
-            // repalce JPA traversible resolver with default implementation (it would be uset automatically
-            // if jpa jar would not appear in classpath
-            bootstrap.setValidatorFactory(Validators.newConfiguration()
-                    .traversableResolver(new TraverseAllResolver())
-                    .buildValidatorFactory());
-        }
-    }
-
-    /**
-     * Hibernate validation hack is requried only if object support is available on classpath.
-     * Without it, real jpa may be used together with graph api.
-     *
-     * @return true if object api detected, false otherwise
-     */
-    private boolean isObjectOrientUsed() {
-        // check whether we have Persistence on the classpath
-        try {
-            Class.forName("com.orientechnologies.orient.object.db.OObjectDatabaseTx",
-                    false, this.getClass().getClassLoader());
-            return true;
-        } catch (ClassNotFoundException e) {
-            return false;
         }
     }
 }