+++ src/main/java/ru/vyarus/dropwizard/orient/configuration/OrientServerConfiguration.java
@@ -10,8 +10,8 @@ import com.orientechnologies.orient.server.config.OServerConfigurationLoaderXml;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Paths;