+++ src/test/groovy/ru/vyarus/dropwizard/orient/support/TestConfiguration.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import io.dropwizard.core.Configuration;
 import ru.vyarus.dropwizard.orient.configuration.OrientServerConfiguration;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 /**
  * Groovy class can't be used for configuraton anymore, because jersey 2.5 is very sensible for addition methods